Rename file to pascal case
