+++ app/routes/admin/email/CreateRestrictedMailRoute.js
@@ -3,7 +3,7 @@
 import { compose } from 'redux';
 import { connect } from 'react-redux';
 import RestrictedMailEditor from './components/RestrictedMailEditor';
-import { createRestrictedMail } from 'app/actions/restrictedMailActions';
+import { createRestrictedMail } from 'app/actions/RestrictedMailActions';
 import { push } from 'react-router-redux';
 
 const mapDispatchToProps = { mutateFunction: createRestrictedMail, push };
